refactor(#112) : 게시글 리펙토링
