+++ src/main/java/com/hobbyhop/domain/post/repository/PostRepository.java
@@ -1,6 +1,7 @@
 package com.hobbyhop.domain.post.repository;
 
 import com.hobbyhop.domain.post.entity.Post;
+import com.hobbyhop.domain.post.repository.custom.PostRepositoryCustom;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long>, PostRepositoryCustom {
