+++ src/main/java/com/hobbyhop/domain/post/service/impl/PostServiceImpl.java
@@ -9,15 +9,13 @@ import com.hobbyhop.domain.post.entity.Post;
 import com.hobbyhop.domain.post.repository.PostRepository;
 import com.hobbyhop.domain.post.s3.S3Service;
 import com.hobbyhop.domain.post.service.PostService;
-import com.hobbyhop.domain.postlike.service.PostLikeService;
+import com.hobbyhop.domain.postuser.service.PostUserService;
 import com.hobbyhop.domain.user.entity.User;
 import com.hobbyhop.global.exception.post.PostNotCorrespondUser;
 import com.hobbyhop.global.exception.post.PostNotFoundException;
 import com.hobbyhop.global.security.userdetails.UserDetailsImpl;
 import java.io.IOException;
-import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -29,7 +27,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class PostServiceImpl implements PostService {
 
     private final ClubService clubService;
-    private final PostLikeService postLikeService;
+    private final PostUserService postUserService;
     private final PostRepository postRepository;
     private final S3Service s3Service;
 
@@ -140,6 +138,6 @@ public class PostServiceImpl implements PostService {
         User user = userDetails.getUser();
         Post post = findAndCheckPostAndClub(clubId, postId);
 
-        postLikeService.postLike(user, post);
+        postUserService.postLike(user, post);
     }
 }
