+++ src/main/java/com/hobbyhop/domain/postuser/service/PostUserService.java
@@ -1,10 +1,9 @@
-package com.hobbyhop.domain.postlike.service;
+package com.hobbyhop.domain.postuser.service;
 
 import com.hobbyhop.domain.post.entity.Post;
-import com.hobbyhop.domain.postlike.entity.PostLike;
 import com.hobbyhop.domain.user.entity.User;
 
-public interface PostLikeService {
+public interface PostUserService {
 
     void postLike(User user, Post post);
 }
